backend/requirements.txt
psutil>=5.9.0
mysql-connector-python>=8.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
joblib>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
# Process list app icons (Windows .exe)
icoextract>=0.2.0
Pillow>=9.0.0
